Move common to api and delete common project
