+++ api/build.gradle.kts
@@ -10,7 +10,6 @@ repositories {
 dependencies {
     kapt(libs.spring.config.processor)
 
-    implementation(project(":common"))
     implementation(libs.bundles.rest.base)
     implementation(libs.spring.mybatis)
     implementation(libs.kulid)
