+++ api/src/main/kotlin/com/example/todo/application/controller/TodoExceptionHandler.kt
@@ -1,6 +1,6 @@
 package com.example.todo.application.controller
 
-import com.example.todo.exceptions.NotFoundException
+import com.example.todo.domain.exception.NotFoundException
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.http.ProblemDetail
