+++ api/src/test/kotlin/com/example/todo/application/controller/TaskControllerTest.kt
@@ -1,11 +1,11 @@
 package com.example.todo.application.controller
 
+import com.example.todo.application.dto.TaskRegistrationDto
+import com.example.todo.application.dto.TaskUpdateDto
 import com.example.todo.domain.entity.TaskEntity
+import com.example.todo.domain.enumration.TaskStatus
+import com.example.todo.domain.exception.NotFoundException
 import com.example.todo.domain.service.TaskService
-import com.example.todo.dto.TaskRegistrationDto
-import com.example.todo.dto.TaskUpdateDto
-import com.example.todo.enums.TaskStatus
-import com.example.todo.exceptions.NotFoundException
 import com.fasterxml.jackson.databind.ObjectMapper
 import io.mockk.MockKAnnotations
 import io.mockk.every
