+++ settings.gradle.kts
@@ -1,3 +1,3 @@
 rootProject.name = "todo"
 
-include("api", "common")
+include("api")
